requirements.txt
 pymongo
 groq
 python-dotenv